Revamp static file handling
