+++ Procfile
@@ -1,2 +1,2 @@
 main: bin/run-haskell
-tailwind: cd content && tailwind-run -w -o static/tailwind.css '../src/**/*.hs' 
+tailwind: cd content && tailwind-run -w -o tailwind.css '../src/**/*.hs'
