+++ ema-template.cabal
@@ -45,6 +45,7 @@ executable ema-template
     , shower
     , some
     , sop-core
+    , generic-optics
     , tagged
     , text
     , time
@@ -105,9 +106,6 @@ executable ema-template
     ViewPatterns
 
   main-is:            Main.hs
-  other-modules:
-    Site.MarkdownSite
-    Site.StaticSite
 
   hs-source-dirs:     src
   default-language:   Haskell2010
